Refactor Mary Mack rhyming game
Added a bit of introductory text as well. I would consider the whole logic behind triggering the scene. Giving Bobby a "current fixation" value of some sort (each pointing to some activity that he does until he's bored), one of them being "wait for player to appear and mess with him", would probably work better for the purposes of the story. Plus, this would give Bobby the ability to vary his way of annoying the player in ways that he thinks are funny but that may actually be deadly dangerous (locking him inside or out of a room, say, or turning the lights out to scare him).